+++ db/migration/src/lib.rs
@@ -13,7 +13,7 @@ mod m20240301_000010_tickets_create_ticket;
 mod m20240301_000011_tickets_create_ticket_stats;
 mod m20240301_000012_tickets_create_outgoing_ticket_index;
 mod m20240404_000013_tickets_recreate_ticket;
-mod m20240810_000013_index_extend_chain_info_with_pre_checksum_block;
+mod m20240810_000014_index_extend_chain_info_with_pre_checksum_block;
 
 #[derive(PartialEq)]
 pub enum BackendType {
@@ -48,7 +48,7 @@ impl MigratorTrait for Migrator {
             Box::new(m20240404_000013_tickets_recreate_ticket::Migration(
                 BackendType::Postgres,
             )),
-            Box::new(m20240810_000013_index_extend_chain_info_with_pre_checksum_block::Migration),
+            Box::new(m20240810_000014_index_extend_chain_info_with_pre_checksum_block::Migration),
         ]
     }
 }
@@ -72,7 +72,7 @@ impl MigratorTrait for MigratorIndex {
             Box::new(m20240226_000006_index_create_network_eligibility::Migration),
             Box::new(m20240226_000008_node_create_settings::Migration),
             Box::new(m20240226_000007_index_initial_seed::Migration),
-            Box::new(m20240810_000013_index_extend_chain_info_with_pre_checksum_block::Migration),
+            Box::new(m20240810_000014_index_extend_chain_info_with_pre_checksum_block::Migration),
         ]
     }
 }
